Add support for partial env destroy.
