+++ cli/aws_orbit/commands/destroy.py
@@ -15,27 +15,44 @@
 import logging
 
 import botocore.exceptions
-from aws_orbit import bundle, plugins, remote
+from aws_orbit import bundle, cleanup, plugins, remote
 from aws_orbit.changeset import Changeset, extract_changeset
 from aws_orbit.manifest import Manifest
 from aws_orbit.messages import MessagesContext
-from aws_orbit.services import cfn, codebuild, s3
+from aws_orbit.services import cfn, codebuild, s3, vpc
 
 _logger: logging.Logger = logging.getLogger(__name__)
 
 
 def destroy_toolkit(manifest: Manifest) -> None:
     if manifest.toolkit_s3_bucket is None:
-        manifest.fillup()
         if manifest.toolkit_s3_bucket is None:
-            _logger.debug("Skipping Toolkit destroy. Because manifest.toolkit_s3_bucket is None")
-            return
+            s3.delete_bucket_by_prefix(
+                manifest=manifest, prefix=f"orbit-{manifest.name}-toolkit-{manifest.account_id}-"
+            )
+        else:
+            try:
+                s3.delete_bucket(manifest=manifest, bucket=manifest.toolkit_s3_bucket)
+            except Exception as ex:
+                _logger.debug("Skipping Toolkit bucket deletion. Cause: %s", ex)
     if cfn.does_stack_exist(manifest=manifest, stack_name=manifest.toolkit_stack_name):
-        try:
-            s3.delete_bucket(manifest=manifest, bucket=manifest.toolkit_s3_bucket)
-        except Exception as ex:
-            _logger.debug("Skipping Toolkit bucket deletion. Cause: %s", ex)
         cfn.destroy_stack(manifest=manifest, stack_name=manifest.toolkit_stack_name)
+
+
+def destroy_remaining_resources(manifest: Manifest) -> None:
+    if cfn.does_stack_exist(manifest=manifest, stack_name=manifest.demo_stack_name):
+        try:
+            vpc_id: str = vpc.get_env_vpc_id(manifest=manifest)
+            cleanup.demo_remaining_dependencies(manifest=manifest, vpc_id=vpc_id)
+        except:  # noqa
+            _logger.debug("VPC not found.")
+        cfn.destroy_stack(manifest=manifest, stack_name=manifest.demo_stack_name)
+    s3.delete_bucket_by_prefix(manifest=manifest, prefix=f"orbit-{manifest.name}-cdk-toolkit-{manifest.account_id}-")
+    env_cdk_toolkit: str = f"orbit-{manifest.name}-cdk-toolkit"
+    if cfn.does_stack_exist(manifest=manifest, stack_name=env_cdk_toolkit):
+        cfn.destroy_stack(manifest=manifest, stack_name=env_cdk_toolkit)
+
+    destroy_toolkit(manifest=manifest)
 
 
 def destroy(env: str, teams_only: bool, keep_demo: bool, debug: bool) -> None:
@@ -43,6 +60,7 @@ def destroy(env: str, teams_only: bool, keep_demo: bool, debug: bool) -> None:
         manifest = Manifest(filename=None, env=env, region=None)
         if manifest.raw_ssm is None:
             ctx.info(f"Environment {env} not found")
+            destroy_remaining_resources(manifest=manifest)
             ctx.progress(100)
             return
